integrate "prompting" experiments into pipeline
